+++ src/rgw/store/sfs/sqlite/versioned_object/versioned_object_conversions.cc
@@ -36,6 +36,7 @@ DBOPVersionedObjectInfo get_rgw_versioned_object(const DBVersionedObject & objec
   rgw_object.size = object.size;
   decode_blob(object.creation_time, rgw_object.creation_time);
   rgw_object.object_state = get_object_state(object.object_state);
+  rgw_object.version_id = object.version_id;
   return rgw_object;
 }
 
@@ -48,6 +49,7 @@ DBVersionedObject get_db_versioned_object(const DBOPVersionedObjectInfo & object
   db_object.size = object.size;
   encode_blob(object.creation_time, db_object.creation_time);
   db_object.object_state = get_uint_object_state(object.object_state);
+  db_object.version_id = object.version_id;
   return db_object;
 }
 }  // namespace rgw::sal::sfs::sqlite